Discord bot setup
This adds a Django management command for starting the bot. This means there will need to be different processes if you want the webpage up at the same time.

This can be beneficial as you will be able to make database changes from both.

Commands are set up through extensions as Cogs which means additional apps can add their own commands as long as they add them to the discord settings.py file