Holidays
This implements holiday models that can be used in other modules.

This is the first part of some changes that will come to source/activity that will allow better controls over when holidays are and what happens on holidays.

This includes Fixed holidays like Christmas or New Year's and Weekday holidays like Thanksgiving.

There are future plans to include holidays based on lunar and planetary cycles. This will allow adding existing holidays like Easter or creating your own lunar and planetary holidays for thematic servers or just to track planetary events.